Refactor topology dictionary generation
It gets rid of a lot of code duplication and unifies the way
codegen.py and gds_dictgen generate topology dictionaries.
